requirements.txt
# Core dependencies
one-api>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning
scikit-learn>=1.0.0

# Utilities
tqdm>=4.60.0
